guestbook_api_proj/requirements.txt
fastapi

# Server
uvicorn

# Postgres
psycopg2


# enviroment
python-dotenv


# Validators & password hashing
email-validator
passlib
bcrypt


python-multipart


